backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
paddleocr>=2.7.0
erniebot
opencv-python-headless